pyproject.toml
[tool.poetry]
name = "slowex-dev"
version = "0.1.0"
description = ""
authors = ["Slava Golyshevskii <sgolyshevskii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
pandas = "^2.2.2"
numpy = "^2.0.1"
apache-airflow = "2.8.3"


[tool.black]
line-length = 88
target-version = ['py39']
exclude = '''
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  '''
skip-string-normalization = false
skip-magic-trailing-comma = true

[tool.isort]
profile = "black"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
